# reports/openapi.yaml
openapi: 3.0.3
info:
  title: Kapda Co. API
  version: 1.0.0
  description: E-commerce platform API for designer marketplace
  contact:
    name: API Support
    email: api@kapdaco.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.kapdaco.com/api
    description: Production server
...
paths:
...
  /public/products:
    get:
      tags: [Public]
      summary: List products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
            enum: [men, women, accessories]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Products list
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /public/products/{id}:
    get:
      tags: [Public]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  product:
                    $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
  /public/designers:
    get:
      tags: [Public]
      summary: List designers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Designers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  designers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Designer'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        mainImage:
          type: string
        stock:
          type: integer
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        material:
          type: string
        isApproved:
          type: boolean
        status:
          type: string
          enum: [draft, pending_review, published]
        createdAt:
          type: string
          format: date-time
...
    Designer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        designerName:
          type: string
        username:
          type: string
        bio:
          type: string
        city:
          type: string
        country:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        productCount:
          type: integer
        followerCount:
          type: integer
        specialties:
          type: array
          items:
            type: string
...
    Pagination:
      type: object
      properties:
        current:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Product not found
...